% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{expect_registry}
\alias{expect_registry}
\title{Does the object match the registry, and should it match?}
\usage{
expect_registry(object, reg, match_expected = TRUE)
}
\arguments{
\item{object}{Character. Text to test}

\item{reg}{Character. Name of registry to test \code{object} against.}

\item{match_expected}{Logical. Whether object should match registry. `TRUE` for true positives. `FALSE` for true negatives.}
}
\description{
Test expectation of whether a TRN come from a particular registry using for known true TRNs and TRN errors.
}
\examples{
\dontrun{expect_registry("PER-093-09", "REPEC", match_expected = TRUE)}
\dontrun{expect_registry("cucumber", "REPEC", match_expected = TRUE)}
\dontrun{expect_registry("cucumber", "REPEC", match_expected = FALSE)}
\dontrun{expect_registry("PER-093-09", "REPEC", match_expected = FALSE)}
}
