% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_trn.R
\name{has_trn}
\alias{has_trn}
\title{Check whether each element of input includes a trial registration number}
\usage{
has_trn(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Boolean, length of input
}
\description{
Check whether each element of input includes a trial registration number
}
\examples{
has_trn(c("NCT00312962", "hello", "euctr2020-001808-42", NA))
}
