% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_trn_registry.R
\name{mutate_trn_registry}
\alias{mutate_trn_registry}
\title{Mutate new column for trial registration number and registry based on specified column}
\usage{
mutate_trn_registry(data, by, .keep = TRUE)
}
\arguments{
\item{data}{A dataframe.}

\item{by}{Column (bare name) in which to search for trn's (and to join by)}

\item{.keep}{(Optional) Keep rows without trn's. Defaults to TRUE.}
}
\value{
A dataframe, with "trn" and "databank" columns
}
\description{
Mutate new column for trial registration number and registry based on specified column
}
\examples{
mutate_trn_registry(sample_trn_df, text)
mutate_trn_registry(sample_trn_df, text, .keep = FALSE)
}
