% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_trn_df}
\alias{sample_trn_df}
\title{Sample text dataframe for extracting trn's and registries}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{id}{id number}
  \item{text}{string containing trial registration number(s)}
  \item{registry_guess}{string containing registry guess for trn}
}
}
\usage{
sample_trn_df
}
\description{
Sample text dataframe for extracting trn's and registries
}
\keyword{datasets}
