% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_registry.R
\name{which_registries}
\alias{which_registries}
\title{Check which registry each element of input is from}
\usage{
which_registries(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector, length of input.
}
\description{
Check which registry each element of input is from
}
\examples{
which_registries(c("NCT00312962", "hello", "euctr2020-001808-42", NA))

df <- dplyr::tibble(trn = c("NCT00312962", "hello", "euctr2020-001808-42", NA))
dplyr::mutate(df, registry = which_registries(trn))

# Does not work for multiple registries in one element
\dontrun{
which_registries(c("NCT00312962 and euctr2020-001808-42", "hello", "euctr2020-001808-42", NA))
}
}
