% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_registry.R
\name{which_registry}
\alias{which_registry}
\title{Check which registry(ies) input is from}
\usage{
which_registry(x)
}
\arguments{
\item{x}{A string (character vector of length 1)}
}
\value{
A string (character vector of length number of registries). Returns NA if no trial registration number detected.
}
\description{
Check which registry(ies) input is from
}
\examples{
which_registry("NCT00312962")
which_registry("NCT00312962 and euctr2020-001808-42")
which_registry("hello")
which_registry(NA)
}
