% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_trn.R
\name{which_trn}
\alias{which_trn}
\title{Extract TRN(s) from input}
\usage{
which_trn(x, registry = NULL, collapse = ";", clean = TRUE)
}
\arguments{
\item{x}{Character vector.}

\item{registry}{Character. Registry of TRN. Default to NULL for all registries.}

\item{collapse}{Character. When a collapse string is given, (by default ";"),
then the return vector will be collapsed to a single string with the chosen
separator string. If collapse is set to "none", the return is a vector with
all extracted TRNs instead.}

\item{clean}{Logical. By default (TRUE), the individual TRNs will also be
cleaned with `trn_clean`. Skip cleaning by setting to FALSE.}
}
\value{
Character. A vector of the same length as the input or
 of length 1, depending on the value of collapse.
 Returns NA if no trial registration numbers detected.
}
\description{
Extract TRN(s) from input
}
\examples{
which_trn("NCT00312962")
which_trn("NCT00312962 and euctr2020-001808-42")
which_trn("NCT00312962 and euctr2020-001808-42", collapse = "none")
which_trn("NCT00312962 and euctr2020-001808-42", registry = "EudraCT")
which_trn("NCT00312962 and euctr2020-001808-42", registry = "ClinicalTrials.gov")
which_trn("NCT 00312962 and euctr2020-001808-42")
which_trn("NCT 00312962 and euctr2020-001808-42", clean = FALSE)
which_trn("hello") # return NA if nothing was detected
}
