% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_trn.R
\name{which_trn}
\alias{which_trn}
\title{Extract TRN(s) from input}
\usage{
which_trn(x)
}
\arguments{
\item{x}{A string (character vector of length 1)}
}
\value{
A string (character vector of length number of registries). Returns NA if no trial registration number detected.
}
\description{
Extract TRN(s) from input
}
\examples{
which_trn("NCT00312962")
which_trn("NCT00312962 and euctr2020-001808-42")
which_trn("hello")
which_trn(NA)
}
