% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_trn.R
\name{which_trn}
\alias{which_trn}
\title{Extract TRN(s) from input}
\usage{
which_trn(x, collapse = ";", clean = TRUE)
}
\arguments{
\item{x}{A string (character vector of length 1)}

\item{collapse}{A string. When a collapse string is given, (by default ";"),
then the return vector will be collapsed to a single string with the chosen
separator string. If collapse is set to "none", the return is a vector with
all extracted TRNs instead.}

\item{clean}{Logical. By default (TRUE), the individual TRNs will also be
cleaned with `trn_clean`. Skip cleaning by setting to FALSE.}
}
\value{
A string (character vector of length number of registries or
 of length 1, depending on value of collapse).
 Returns NA if no trial registration numbers detected.
}
\description{
Extract TRN(s) from input
}
\examples{
which_trn("NCT00312962")
which_trn("NCT00312962 and euctr2020-001808-42")
which_trn("NCT00312962 and euctr2020-001808-42", collapse = "none")
which_trn("NCT 00312962 and euctr2020-001808-42")
which_trn("NCT 00312962 and euctr2020-001808-42", clean = FALSE)
which_trn("hello") # return NA if nothing was detected
which_trn(c("NCT 00312962", NA, "", "hello")) # NAs and empty strings are ignored

}
