% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_trn.R
\name{which_trn}
\alias{which_trn}
\title{Extract TRN(s) from input}
\usage{
which_trn(x, collapse = NA_character_)
}
\arguments{
\item{x}{A string (character vector of length 1)}

\item{collapse}{A string. By default, NA, the return is a vector with
all extracted TRNs. If a collapse string is given, then the vector will
be collapsed to a single string with the chosen separator string.}
}
\value{
A string (character vector of length number of registries or
 of length 1, depending on value of collapse).
 Returns NA if no trial registration numbers detected.
}
\description{
Extract TRN(s) from input
}
\examples{
which_trn("NCT00312962")
which_trn("NCT00312962 and euctr2020-001808-42")
which_trn("NCT00312962 and euctr2020-001808-42", collapse = ";")
which_trn("hello")
which_trn(NA, NA)
}
