% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_trn.R
\name{which_trns}
\alias{which_trns}
\title{Extract TRN(s) from each element of input}
\usage{
which_trns(trn_vec, collapse = NA_character_, clean = TRUE)
}
\arguments{
\item{trn_vec}{A character vector.}

\item{collapse}{A string. By default, NA, the return is a vector with
all extracted TRNs. If a collapse string is given, then the vector will
be collapsed to a single string with the chosen separator string.}

\item{clean}{Logical. By default (TRUE), the individual TRNs will also be
cleaned with `trn_clean`. Skip cleaning by setting to FALSE.}
}
\value{
A character vector, length of input.
}
\description{
Extract TRN(s) from each element of input
}
\examples{
which_trns(c("NCT00312962", "hello", "euctr2020-001808-42", NA))

df <- dplyr::tibble(trn = c("NCT00312962", "hello", "euctr2020-001808-42", NA))
dplyr::mutate(df, trn_extract = which_trns(trn))

# docDoes not work for multiple TRNs in one element if no collapse chosen
\dontrun{
which_trns(c("NCT00312962 and euctr2020-001808-41", "hello", "euctr2020-001808-42", NA))
}
which_trns(c("NCT 00312962 and euctr2020-001808-41", "hello", "euctr2020-001808-42", NA),
 collapse = ";")
which_trns(c("NCT 00312962 and euctr2020-001808-41", "hello", "euctr2020-001808-42", NA),
 collapse = ";", clean = FALSE)
}
